service1/requirements.txt
pika==1.3.2
aio-pika==9.3.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation==0.44b0
pydantic-settings==2.2.1
